Fix IO bugs, UI, internal issues
Switched to QFile and QDataStream instead of std::ifstream to enable proper use of strings.
FIxes IO behaviours.
Fixed UI bugs while loading with ports.
Added color alpha in the picker.
Optimise parts of code especially nodeconnector.
Using platform independent, fixed size types for data file storing.
Using coherent ID types in both UI and file data types.
Added OperationalMutex to avoid reading while writing (generating UB).

+++ connectionitem.cpp
@@ -22,7 +22,7 @@ ConnectionItem::ConnectionItem(QGraphicsItem *parent) : QGraphicsPathItem(parent
 
 ConnectionItem::~ConnectionItem() {}
 
-int ConnectionItem::UID()
+qint64 ConnectionItem::UID()
 {
     return m_uid;
 }
